requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

pydantic>=2.6.0
python-dotenv>=1.0.1

passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

asyncpg>=0.29.0

slowapi>=0.1.9
limits>=3.9.0

email-validator>=2.1.0
